// packages/lodestar/src/chain/blocks/types.ts
import {CachedBeaconStateAllForks} from "@chainsafe/lodestar-beacon-state-transition";
import {ExecutionStatus} from "@chainsafe/lodestar-fork-choice";
import {allForks, Slot} from "@chainsafe/lodestar-types";

export type ImportBlockOpts = {
  /**
   * TEMP: Review if this is safe, Lighthouse always imports attestations even in finalized sync.
   */
  skipImportingAttestations?: boolean;
  /**
   * If error would trigger BlockErrorCode ALREADY_KNOWN or GENESIS_BLOCK, just ignore the block and don't verify nor
   * import the block and return void | Promise<void>.
   * Used by range sync and unknown block sync.
   */
  ignoreIfKnown?: boolean;
  /**
   * If error would trigger WOULD_REVERT_FINALIZED_SLOT, it means the block is finalized and we could ignore the block.
   * Don't import and return void | Promise<void>
   * Used by range sync.
   */
  ignoreIfFinalized?: boolean;
  /**
   * From RangeSync module, we won't attest to this block so it's okay to ignore a SYNCING message from execution layer
   */
  fromRangeSync?: boolean;
  /**
   * Verify signatures on main thread or not.
   */
  blsVerifyOnMainThread?: boolean;
  /**
   * Metadata: `true` if only the block proposer signature has been verified
   */
  validProposerSignature?: boolean;
  /**
   * Metadata: `true` if all the signatures including the proposer signature have been verified
   */
  validSignatures?: boolean;
};

/**
 * A wrapper around a `SignedBeaconBlock` that indicates that this block is fully verified and ready to import
 */
export type FullyVerifiedBlock = {
  block: allForks.SignedBeaconBlock;
  postState: CachedBeaconStateAllForks;
  parentBlockSlot: Slot;
  /**
   * If the execution payload couldnt be verified because of EL syncing status, used in optimistic sync or for merge block
   */
  executionStatus: ExecutionStatus;
};